+++ SourceryFramework/Sources/Parsing/SwiftSyntax/FileParserSyntax.swift
@@ -1,5 +1,6 @@
 import Foundation
 import SwiftSyntax
+import SwiftSyntaxParser
 import PathKit
 import SourceryRuntime
 import SourceryUtils